+++ Assets/Prefabs/CardPrefabs/CardCrushingBlow.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: eb774535e23131b429e4217dc5d9ca33
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
