+++ Assets/Prefabs/Deck Temp.prefab
@@ -94,13 +94,13 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: ffa7d28e7d81cbb4a86e58ccadcf2e74, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  card1: {fileID: 3383020702869181935, guid: 2916b68f5b43dc541b34c63054fcd62b, type: 3}
-  card2: {fileID: 6805844535874794075, guid: 6bb712162cec5a24e8301b1a635bdd50, type: 3}
-  card3: {fileID: 2251936837461568767, guid: 8521770a979de56499ae7577399f47e0, type: 3}
+  card1: {fileID: 7976870046634001099, guid: 54cb9d269bc86104c981de9294a6e904, type: 3}
+  card2: {fileID: 3383020702869181935, guid: 2916b68f5b43dc541b34c63054fcd62b, type: 3}
+  card3: {fileID: 497074487205763524, guid: 0ac4cb796f65ec1468260bf123619042, type: 3}
   card4: {fileID: 2769199774944758380, guid: f50c91d7b0762d5438f94372c742b825, type: 3}
   card5: {fileID: 7886511964787177903, guid: 1181f82ee8a22a044a21b7f97ed7af76, type: 3}
-  card6: {fileID: 6459855802533077854, guid: b3215e5c54d41cf45956e6cf9d0c8093, type: 3}
+  card6: {fileID: 2251936837461568767, guid: 8521770a979de56499ae7577399f47e0, type: 3}
   card7: {fileID: 4542706657116882137, guid: e01944f53b1640547bdbffac548b4233, type: 3}
   card8: {fileID: 3131678796359061473, guid: b11c0eabb72faad4c962d0ea9c26733b, type: 3}
-  card9: {fileID: 4014162451266605590, guid: e66247aaeee8704459ad9e6f8c2d5e21, type: 3}
-  card10: {fileID: 3185659527727028189, guid: 78b7e96ee3e6d464986d5a216f7b9f20, type: 3}
+  card9: {fileID: 1659685387865724933, guid: eb774535e23131b429e4217dc5d9ca33, type: 3}
+  card10: {fileID: 6805844535874794075, guid: 6bb712162cec5a24e8301b1a635bdd50, type: 3}
