--- Assets/Scripts/Cards/Card10.cs
@@ -1,21 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-public class Card10 : Card
-{
-    // Start is called before the first frame update
-    void Start()
-    {
-        cardName = "Card10";
-        id = 10;
-        mana = 6;
-        damage = 35; //will be for two turns (total 70)
-    }
-
-    // Update is called once per frame
-    void Update()
-    {
-
-    }
-}
