--- Assets/Scripts/Cards/Card10.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 1482e362ca4255f4085d0399a08f6abe
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
