--- Assets/Scripts/Cards/Card2.cs
@@ -1,21 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-public class Card2 : Card
-{
-    void Start()
-    {
-        
-        cardName = "Card2";
-        id = 2;
-        mana = 1;
-        damage = 20;
-    }
-
-    // Update is called once per frame
-    void Update()
-    {
-
-    }
-}
