--- Assets/Scripts/Cards/Card2.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 75faa913b932af34a95939bf753b6245
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
