--- Assets/Scripts/Cards/Card3.cs
@@ -1,21 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-public class Card3 : Card
-{
-    // Start is called before the first frame update
-    void Start()
-    {
-        cardName = "Card3";
-        id = 3;
-        mana = 2;
-        damage = -15;
-    }
-
-    // Update is called once per frame
-    void Update()
-    {
-
-    }
-}
