--- Assets/Scripts/Cards/Card6.cs
@@ -1,21 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-public class Card6 : Card
-{
-    // Start is called before the first frame update
-    void Start()
-    {
-        cardName = "Card6";
-        id = 6;
-        mana = 3;
-        damage = 40;
-    }
-
-    // Update is called once per frame
-    void Update()
-    {
-
-    }
-}
