--- Assets/Scripts/Cards/Card6.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 4441d84cd2aafbe4db136eac6258e534
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
