--- Assets/Scripts/Cards/Card9.cs
@@ -1,21 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-public class Card9 : Card
-{
-    // Start is called before the first frame update
-    void Start()
-    {
-        cardName = "Card9";
-        id = 9;
-        mana = 5;
-        damage = 55;
-    }
-
-    // Update is called once per frame
-    void Update()
-    {
-
-    }
-}
