+++ Assets/Scripts/Cards/CardCrushingBlow.cs
@@ -2,25 +2,27 @@ using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
 
-public class CardAttack3 : Card
+public class CardCrushingBlow : Card
 {
     public List<Enemy> Targets;
-    
-    
+
+
     // Start is called before the first frame update
     void Start()
     {
-        mana = 3;
-        numberOfTargets = 3;
+        id = 9;
+        damage = 55;
+        mana = 5;
+        numberOfTargets = 1;
         Targeter = this.gameObject.GetComponent<SelectionGO>();
         Targeter.numberOfSelections = numberOfTargets;
-        Targeter.exclusive = false;
+        Targeter.exclusive = true;
     }
 
     // Update is called once per frame
     void Update()
     {
-        
+
     }
 
     override public void Action()
@@ -29,7 +31,7 @@ public class CardAttack3 : Card
         {
             Enemy e = GO.GetComponent<Enemy>();
             if (e != null)
-                e.TakeDamage(20);
+                e.TakeDamage(damage);
         }
         RemoveHighlightTargets();
         ClearSelections();
